Reverting r1124.
This change was originally made to allow a client to continue without error even
if the response handler did not completely read the response. This doesn't make 
a lot of sense, and indicates an error in the implementation of the response 
handler. 
Besides that, it makes serf abort (double free) when the response handler has 
changed the request queue.