Serf combines the values of two headers with the same name into one, separated by a comma. That's nice and all, but makes it very difficult to parse such values, especially in the case of WWW-Authenticate headers where the values contain lists of comma separated arguments.
This patch moves the header combining code from the header-set function
to the header-get function. The only difference is that we can now
use serf_bucket_headers_do to get multiple headers with the same name
separately, and get them in the same order as they were added (which
is important, because we want e.g. if the server supports NTLM and 
Basic authentication, the NTLM header will come first and has to be 
tried first.

* buckets/headers_buckets.c
  (serf_bucket_headers_setx): Don't combine headers with the same name
   here, just add them at the end of the list.
  (serf_bucket_headers_get): Do the combining here.

* test/test_buckets.c
  (test_bucket_header_set): add some small test cases.

* test/test_context.c
  (test_serf_setup_proxy): Headers are now sent on the socket in the order
   in which they were set.