Merge the branch fix-authn-errors to trunk.
This branch fixes the problem of reporting response errors to the application
when an authentication callback handler is set.

A first attempt to fix it was made in r2360, but with some bad side effects.
The approach implemented in this branch is explained in this mail thread
(and in the code comments):
https://groups.google.com/forum/#!msg/serf-dev/3_HLwF2lVsY/bcR-fatxGB0J